clean up

+++ sql/Weather.sql
@@ -1,8 +1,8 @@
 Create table Weather (
     Location VARCHAR(20),
     DateStamp DATE,
-    MaxTemp int,
-    MinTemp int,
-    Precipitation int,
+    MaxTemp float,
+    MinTemp float,
+    Precipitation float,
     CONSTRAINT PK_WeatherEntry PRIMARY KEY (Location, DateStamp)
 );
